# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmUI API
  description: |
    API documentation for SwarmUI - AI-powered development environment with parallel CLI agent orchestration.
    
    ## Overview
    SwarmUI provides a comprehensive API for managing:
    - Chat sessions with AI agents
    - Project and ticket management
    - Background job processing
    - Git operations
    - File system operations
    - Terminal sessions
    - Testing and debugging
    
    ## Authentication
    Most endpoints require authentication via NextAuth.js session cookies.
    
    ## WebSocket
    Real-time updates are available via WebSocket at `ws://localhost:3000`.
    See [WebSocket API documentation](./WEBSOCKET_API.md) for details.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SwarmUI Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:4000
    description: Alternative dev port
...
paths:
...
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
...
  /api/git/status:
    get:
      tags: [Git]
      summary: Get git status
      operationId: getGitStatus
      parameters:
        - name: path
          in: query
          schema:
            type: string
          description: Project path (defaults to settings.projectPath)
      responses:
        '200':
          description: Git status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatus'
...
  /api/mcp:
    post:
      tags: [MCP]
      summary: Execute MCP tool
      operationId: executeMcpTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPToolCall'
      responses:
        '200':
          description: Tool result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPToolResult'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    AgentRole:
      type: string
      enum: [researcher, planner, coder, validator, security, synthesizer]
    CLIProvider:
      type: string
      enum: [cursor, gemini, claude, copilot, codex, rovo, custom]
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        prd:
          type: string
        prdStatus:
          type: string
          enum: [draft, approved, rejected]
        features:
          type: array
          items:
            type: string
        epics:
          type: array
          items:
            $ref: '#/components/schemas/Epic'
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
    Epic:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        ticketIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, completed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        complexity:
          $ref: '#/components/schemas/TicketComplexity'
        status:
          $ref: '#/components/schemas/TicketStatus'
        level:
          $ref: '#/components/schemas/TicketLevel'
        assignedRole:
          $ref: '#/components/schemas/AgentRole'
        assignedProvider:
          $ref: '#/components/schemas/CLIProvider'
        parentId:
          type: string
        epicId:
          type: string
        dependencies:
          type: array
          items:
            type: string
        output:
          type: string
        diff:
          type: string
        confidence:
          type: number
        sla:
          $ref: '#/components/schemas/SLAConfig'
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    TicketComplexity:
      type: string
      enum: [S, M, L, XL]
    TicketStatus:
      type: string
      enum: [backlog, in_progress, review, approved, rejected, done]
    TicketLevel:
      type: string
      enum: [feature, epic, story, task, subtask, subatomic]
    SLAConfig:
      type: object
      properties:
        responseTimeHours:
          type: number
        resolutionTimeHours:
          type: number
        priority:
          type: string
          enum: [low, medium, high, critical]
...
    GitStatus:
      type: object
      properties:
        branch:
          type: string
        upstream:
          type: string
        ahead:
          type: integer
        behind:
          type: integer
        staged:
          type: array
          items:
            type: string
        unstaged:
          type: array
          items:
            type: string
        untracked:
          type: array
          items:
            type: string
...
    MCPToolCall:
      type: object
      required: [serverId, toolName]
      properties:
        serverId:
          type: string
        toolName:
          type: string
        args:
          type: object
    MCPToolResult:
      type: object
      properties:
        serverId:
          type: string
        toolName:
          type: string
        result:
          type: object
        error:
          type: string
        timestamp:
          type: integer
...